refactor(runtime): return provider credentials as spec entries and drop redundant hasIdentity
