+++ src/assets/templates/agent-typescript-strands/README.md
@@ -22,7 +22,7 @@ validation when extending the request shape, and pass only prompt text to the ag
 
 | Variable | Required | Description |
 | --- | --- | --- |
-{{#if hasIdentity}}| `{{identityProviders.[0].envVarName}}` | Yes | {{modelProvider}} API key (local) or Identity provider name (deployed) |
+{{#if identityProviders.[0]}}| `{{identityProviders.[0].envVarName}}` | Yes | {{modelProvider}} API key (local) or Identity provider name (deployed) |
 {{/if}}| `LOCAL_DEV` | No | Set to `1` to use `.env.local` instead of AgentCore Identity |
 
 # Developing locally
