+++ src/core/project/templates/types.ts
@@ -16,6 +16,21 @@ export type SpecEntries = {
   evaluators?: Evaluator[];
 };
 
+/** Combines several {@link SpecEntries} into one, concatenating each resource collection. */
+export function mergeSpecEntries(entries: SpecEntries[]): SpecEntries {
+  const runtimes = entries.flatMap(({ runtimes }) => runtimes ?? []);
+  const credentials = entries.flatMap(({ credentials }) => credentials ?? []);
+  const memories = entries.flatMap(({ memories }) => memories ?? []);
+  const harnesses = entries.flatMap(({ harnesses }) => harnesses ?? []);
+
+  return {
+    ...(runtimes.length > 0 && { runtimes }),
+    ...(credentials.length > 0 && { credentials }),
+    ...(memories.length > 0 && { memories }),
+    ...(harnesses.length > 0 && { harnesses }),
+  };
+}
+
 /** A group of files and resources that can be rendered into a project **/
 export type Template = {
   tree: FsTreeNode;
